src/main/java/com/example/securingapis2/security/JwtTokenUtil.java
package com.example.securingapis2.security;

import io.jsonwebtoken.Claims;
import io.jsonwebtoken.Jwts;
import io.jsonwebtoken.SignatureAlgorithm;
import io.jsonwebtoken.security.Keys;
import org.springframework.stereotype.Component;

import javax.crypto.SecretKey;
import java.util.Date;

@Component
public class JwtTokenUtil {

    // Define a secret key for signing the JWT
    private final SecretKey secretKey = Keys.secretKeyFor(SignatureAlgorithm.HS256); // Predefined key for signing

    // Generate the signing key for JWT
    public SecretKey getSigningKey() {
        return secretKey;  // Can be used in other methods for signing tokens
    }

    // Method to generate token
    public String generateToken(String username) {
        // Set token expiration time (e.g., 10 minutes)
        long expirationTime = 10 * 60 * 1000;

        return Jwts.builder()
                .setSubject(username)
                .setIssuedAt(new Date())
                .setExpiration(new Date(System.currentTimeMillis() + expirationTime))
                .signWith(getSigningKey())  // Use the method to get the signing key
                .compact();
    }

    // Method to validate token
    public boolean validateToken(String token, String username) {
        String tokenUsername = getUsernameFromToken(token);
        return (tokenUsername.equals(username) && !isTokenExpired(token));
    }

    // Extract username from token
    public String getUsernameFromToken(String token) {
        Claims claims = getAllClaimsFromToken(token);
        return claims.getSubject();
    }

    // Extract all claims from the token
    private Claims getAllClaimsFromToken(String token) {
        return Jwts.parserBuilder()
                .setSigningKey(getSigningKey())  // Use the method to get the signing key
                .build()
                .parseClaimsJws(token)
                .getBody();
    }

    // Check if the token has expired
    private boolean isTokenExpired(String token) {
        Date expiration = getAllClaimsFromToken(token).getExpiration();
        return expiration.before(new Date());
    }
}
